Add dstringbuilder, which greatly decreases memory usage.
In the case where no files are updated, Spark now uses less than
half the memory it did pre-dstringbuilder. Even if all files
need to be updated, I believe it will still consume less memory
than it did previously.

There was a lot of excess duplication of certain data in the
html_page_creators, specifically regarding post page content.
dstringbuilder stores dstrings and dstringbuilders,
and is perfectly suited for surrounding a dstring with more
dstrings on either side (as is the case when eg a post
is wrapped in the post header and footer, which is wrapped
in the <main> tag, which is wrapped in the page header and
footer), without wasting space.

There are dstringbuilder functions related to reading/writing
files, so that the dstringbuilder doesn't have to be formed
(turned into a single dstring) if there's no change.